Behaviors now take options in the constructor.
Rather than passing in options on each invocation of a Behavior's "serve" method, these are now set in the object's constructor (if there are any defaults) and used as instance variables.  This means that the BehaviorServer no longer needs to accept options as an optional argument, which would then get used for no other purpose of being passed through to the collaborating behavior object.  This removes the "Middle Man" code smell.