Creating an internal object for constructing configurations.
I'm still not clear if this object, LinearPortConfigurationFactory, should be an object internal, or a peer.  Right now I'm leaving it as a peer, which means the tests for CommandLineConfiguration do not change.  If this were to become a dependency, I'd change the tests to pass in a test double, then test LinearPortConfigurationFactory directly in a separate test.