Test that a socket connection is refused.
The previous version used the 'timeout' gem to ensure this never connected.  However, this test passed on Mac OS X 10.8.5 and immediately failed with an Errno::ECONNREFUSED on Ubuntu.  After removing this timeout, I also get an ECONNREFUSED error on the Mac after 8 seconds.

This first step should make the test pass on both platforms.  The next step may be to shrink the timeout on this test so it's not so slow on the Mac - possibly using another timeout technique.